Extracting application middleware
